Add config for old Mattock and Kama behavior
