+++ src/main/java/slimeknights/tconstruct/tools/tools/Kama.java
@@ -34,6 +34,7 @@ import java.util.Random;
 
 import javax.annotation.Nonnull;
 
+import slimeknights.tconstruct.common.config.Config;
 import slimeknights.tconstruct.library.events.TinkerToolEvent;
 import slimeknights.tconstruct.library.materials.Material;
 import slimeknights.tconstruct.library.tinkering.Category;
@@ -97,41 +98,11 @@ public class Kama extends AoeToolCore {
   @Nonnull
   @Override
   public EnumActionResult onItemUse(EntityPlayer player, World world, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
-    ItemStack stack = player.getHeldItem(hand);
-    if(ToolHelper.isBroken(stack)) {
-      return EnumActionResult.FAIL;
+    if (Config.oldMattockAndKama) {
+      return EnumActionResult.PASS;
     }
 
-    EnumActionResult ret = useHoe(stack, player, world, pos, hand, facing, hitX, hitY, hitZ);
-    for(BlockPos blockPos : getAOEBlocks(stack, world, player, pos)) {
-      if(ToolHelper.isBroken(stack)) {
-        break;
-      }
-
-      EnumActionResult ret2 = useHoe(stack, player, world, blockPos, hand, facing, hitX, hitY, hitZ);
-      if(ret != EnumActionResult.SUCCESS) {
-        ret = ret2;
-      }
-    }
-
-    if(ret == EnumActionResult.SUCCESS) {
-      TinkerToolEvent.OnMattockHoe.fireEvent(stack, player, world, pos);
-    }
-
-    return ret;
-  }
-
-  private EnumActionResult useHoe(ItemStack stack, EntityPlayer player, World world, BlockPos blockPos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
-    // make sure no damage is taken
-    int damage = stack.getItemDamage();
-    EnumActionResult ret = Items.DIAMOND_HOE.onItemUse(player, world, blockPos, hand, facing, hitX, hitY, hitZ);
-    stack.setItemDamage(damage);
-
-    // do tinkers damaging
-    if(!world.isRemote && ret == EnumActionResult.SUCCESS) {
-      ToolHelper.damageTool(stack, 1, player);
-    }
-    return ret;
+    return doTill(player, world, pos, hand, facing, hitX, hitY, hitZ);
   }
 
   @Nonnull
